+++ src/datasource/TaskDatasource.ts
@@ -2,6 +2,7 @@ import { BASE_URL, httpClient } from "../config/develop-config";
 import { FarmerEntity } from "../entities/FarmerEntities";
 import {
   CreateNewTaskEntity,
+  GetNewTaskEntity,
   NewTaskPageEntity,
 } from "../entities/NewTaskEntities";
 
@@ -50,5 +51,36 @@ export class TaskDatasource {
       .catch((error) => {
         console.log(error);
       });
+  }
+  static getNewTaskById(id: string): Promise<GetNewTaskEntity> {
+    return httpClient
+      .get(BASE_URL + "/tasks/task/" + id)
+      .then((response) => {
+        return response.data.data;
+      })
+      .catch((err) => {
+        console.log(err, "err getnewtaskbyid");
+      });
+  }
+  static updateNewTask(data: GetNewTaskEntity): Promise<any> {
+    delete data["taskDronerTemp"];
+    return httpClient
+      .patch(BASE_URL + "/tasks/task/" + data.id, data)
+      .then((response) => {
+        return response.data;
+      })
+      .catch((error) => {
+        console.log(error);
+      });
+  }
+  static deleteTask(id:string) :Promise<any>{
+    return httpClient
+    .delete(BASE_URL + "/tasks/task/" + id)
+    .then((response) => {
+      return response.data;
+    })
+    .catch((error) => {
+      console.log(error);
+    });
   }
 }