+++ src/entities/TaskDronerTemp.ts
@@ -6,8 +6,9 @@ export interface TaskDronerTempEntity {
   distance: string;
   createdAt: string;
   dronerDetail: string;
+  isChecked: boolean;
 }
-export const TaskDronerTempEntity_INIT : TaskDronerTempEntity = {
+export const TaskDronerTempEntity_INIT: TaskDronerTempEntity = {
   id: "",
   taskId: "",
   dronerId: "",
@@ -15,14 +16,25 @@ export const TaskDronerTempEntity_INIT : TaskDronerTempEntity = {
   distance: "",
   createdAt: "",
   dronerDetail: "",
-}
+  isChecked: false,
+};
 export interface CreateDronerTempEntity {
+  taskId?: string;
   dronerId: string;
   status: string;
   dronerDetail: string[];
 }
 export const CreateDronerTempEntity_INIT: CreateDronerTempEntity = {
+  taskId: "",
   dronerId: "",
   status: "",
   dronerDetail: [""],
 };
+export interface DeletedDronerTemp {
+  taskId: string;
+  dronerId: string;
+}
+export const DeletedDronerTemp_INIT : DeletedDronerTemp = {
+  taskId: "",
+  dronerId: "",
+}