+++ src/pages/task/newTask/IndexNewTask.tsx
@@ -8,6 +8,7 @@ import { Badge, Button, DatePicker, Dropdown, Menu, Select, Table } from "antd";
 import Search from "antd/lib/input/Search";
 import moment from "moment";
 import React, { useEffect, useState } from "react";
+import Swal from "sweetalert2";
 import ActionButton from "../../../components/button/ActionButton";
 import { CardContainer } from "../../../components/card/CardContainer";
 import Layouts from "../../../components/layout/Layout";
@@ -83,6 +84,23 @@ const IndexNewTask = () => {
     setShowModalDroner((prev) => !prev);
     setTaskId(taskId);
   };
+  const removeNewTask = async (e: string) => {
+    Swal.fire({
+      title: "ยืนยันการลบ",
+      text: "โปรดตรวจสอบงานใหม่ที่คุณต้องการลบ ก่อนที่จะกดยืนยันการลบ เพราะอาจส่งผลต่อการจ้างงานในระบบ",
+      cancelButtonText: "ยกเลิก",
+      confirmButtonText: "ลบ",
+      confirmButtonColor: "#d33",
+      showCancelButton: true,
+      showCloseButton: true,
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        await TaskDatasource.deleteTask(e).then();
+      }
+      fetchNewTaskList();
+    });
+  };
+
   const menu = (
     <Menu
       items={[
@@ -233,7 +251,7 @@ const IndexNewTask = () => {
               </span>
               <br />
               <span style={{ color: color.Grey }}>
-                จำนวน {row.count_droner} ไร่
+                จำนวน {row.farm_area_amount} ไร่
               </span>
             </>
           ),
@@ -304,7 +322,11 @@ const IndexNewTask = () => {
                 />
               </div>
               <div className="col-lg-6">
-                <ActionButton icon={<DeleteOutlined />} color={color.Error} />
+                <ActionButton
+                  icon={<DeleteOutlined />}
+                  color={color.Error}
+                  onClick={() => removeNewTask(row.id)}
+                />
               </div>
             </div>
           ),