+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ## [Version - 0]
+## Fix [v0.14.1] 2022-09-22
+- fixed defect index new task
+- fixed birthDate and format in farmer & droner
+## Feature [v0.14.0] 2022-09-22
+- edit new task
+- selected droner
+- enhanced index new task
+- enhanced status receive task for droner
+- removed new task
+- click botton navigate to google map in modal plot
 ## Fix [v0.13.3] 2022-09-21
 - fixed sub status in index finish task
 - fixed separate data with commas.